Fixed naming scheme
The naming scheme of some graphs are tricky in the way that they not sufixed
by a single numeric digit, but a double with padding (mostly a zero). This
commit fixes the numeric scheme of 'fs' and 'net' graphs, in the event of
when the user selects all graphs of all remote servers.

Still there might be possible that some graph does not appear because the
master server has a different configuration that the remote servers.

#216