Improved the alerting system in 'gensens.pm'
From now on, the alerts accepts a forth parameter that specifies
when the alert will be triggered. This parameter accepts two values:
'above' or 'below', being the 'above' the default value if it's not
specified, in order to keep backwards compatibility.

This way, if some value in the graph is 'above' or 'below' from the
threshold during more time than specified in the timeout value, the
alert will be triggered.

#221