Added a new option called 'autocheck_responsiveness' for the HTTP built-in
There is a well known problem with the HTTP built-in server that Monitorix
implements using the Perl module 'HTTP::Server::Simple'. It looks like it's
pretty weak to the common attacks that any web server receives every day.

As a consequence of that, it just hangs, and when you want to see the stats of
your server, you just see your browser waiting for a response that never comes,
which ends up with a timeout. In these cases, you are forced to login to your
server and restart Monitorix, in order to refresh the HTTP built-in server.

I was frustrated because I was completely unable to reproduce locally the kind
of attack that seems to do a DoS to the HTTP built-in server, so in order to
mitigate (hopefully) all this annoying hangups, I've created an autocheck to
control the responsiveness of the HTTP server on every minute, and in case of
no response then it will be restarted automatically. This is controlled by a new
option called 'autocheck_responsiveness' which by default is enabled.

Let's see how this is working!