Introduced the option 'cmd' in 'netstat.pm'
From now on the user will have the opportunity to select which
command will use to gather the network status statistics. The
new option 'cmd' supports two values 'netstat' or 'ss', being
the later the value by default.

If the 'cmd' option is not defined, then the command line used
will be 'ss', that's because in newer Linux systems it looks like
the command 'netstat' is being deprecated.

In older Linux systems people will be forced to use the 'netstat'
command because, early 'ss' versions doesn't contain the protocol
column in its output, which breaks the regexp used to extract the
the values. #196