Fixed and improved the calculation of 'Used' in .memory.pm'
There was a stupid bug that prevented to include the SReclaimable value in the
formula to calculate the 'Used' value.

Morever, I've seen that newer versions of 'free' command also include the value
SUnreclaim in that formula. So, the final calculation of 'Used' is like this:

Used = MemTotal - MemFree - Buffers - Cached - SReclaimable - SUnreclaim

This ensures that Monitorix is in sync with the results of 'free' command. #204