New option 'default_option_when_all' for Multihost mode
Now that in Multihost mode anyone can select the option "All" in the Hostname list and "All graphs" in the Graph list, this may, accidentally, hang the browser for a while due to the huge amount of images to download remotely from different servers.

In order to prevent precisely that, this new option defines which option in the Graph list will be selected ("System load" by default) automatically when the user selects 'All' in the Hostname list.

Of course, the user is still able to change it to "All graphs" at any moment, and at his own risk :-). #216