Fixed the behavior of the options in the extra configuration file.
A configuration file can contain sections, subsections and even
subsubsections in the form of:

<gensens>	<-- this is a section or module
	<list>		<-- this is a subsection
		0 = temp0
		1 = cpu0
		2 = bat0
	</list>
	...

If you plan to create an extra configuration file in 'conf.d/', you
must know that, from now on, you'll be forced to define the whole
subsection even when you are changing only a single value of it.
Otherwise your results might vary from what you expect.

For instance, imagine you want to monitor a second CPU in the
'gensens' module. You will create an extra configuration file in
'conf.d/' to include a second CPU (cpu1) in the key '1' inside the
subsection 'list'. In that case your extra configuration file
should define the section 'gensens' as this:

<gensens>
	<list>
		0 = temp0
		1 = cpu0, cpu1
		2 = bat0
	</list>
</gensens>

That is, you must include the whole subsection 'list' with all the
rest of options. The rest of subsections of 'gensens' are not needed
because they have not been changed. Monitorix will use the subsections
configured in the main configuration file.