Add additional_graph_name config option.
This is an optional configuration option for configuration files
in include_dir. It takes a comma separated list of additional
modules that is appended to the graph_name option. This enables
you to add custom modules without changing the monitorix.conf or
copy all names from graph_name to your config file.